Cleaned the data frame into to be able to call gmaps using jupyter. Added in screeenshots of the gmaps created from the new data

+++ data_cleanse/city_long_lat.py
@@ -2,28 +2,31 @@ import requests
 import json
 from config import google_places_key
 import pandas as pd
+import os
 
-data = pd.read_csv('../resources/glassdoor_data_scientist_jobs.csv')
-sorted_cities = data["Location"].unique()
+if not os.path.exists('../resources/map_information'):
+    data = pd.read_csv('../resources/glassdoor_data_scientist_jobs.csv')
+    sorted_cities = data["Location"].unique()
 
-url = 'https://maps.googleapis.com/maps/api/geocode/json?address='
-cities = []
-state = []
-long = []
-lat = []
-for x in sorted_cities:
-    state.append(x[-2:])
-    cities.append(x[:-4])
-    payload = requests.get(f'{url}{str(x).replace(" ", "+")}&key={google_places_key}').json()
-    long.append(payload['results'][0]['geometry']['location']['lng'])
-    lat.append(payload['results'][0]['geometry']['location']['lat'])
-    print(f'{x} added!')
+    url = 'https://maps.googleapis.com/maps/api/geocode/json?address='
+    cities = []
+    state = []
+    long = []
+    lat = []
+    for x in sorted_cities:
+        state.append(x[-2:])
+        cities.append(x[:-4])
+        payload = requests.get(f'{url}{str(x).replace(" ", "+")}&key={google_places_key}').json()
+        long.append(payload['results'][0]['geometry']['location']['lng'])
+        lat.append(payload['results'][0]['geometry']['location']['lat'])
+        print(f'{x} added!')
 
-map_info = pd.DataFrame({
-    "city": cities,
-    "state": state,
-    "long": long,
-    "lat": lat
-})
-map_info.to_csv("resources/map_information", index=False)
+    map_info = pd.DataFrame({
+        "city": cities,
+        "state": state,
+        "long": long,
+        "lat": lat
+    })
+    map_info.to_csv("resources/map_information", index=False)
+data = pd.read_csv('../resources/map_information')
